Implements feedback [skip ci]
