--- src/main/java/tdl/client/runner/IImplementationRunner.java
@@ -1,5 +1,0 @@
-package tdl.client.runner;
-
-public interface IImplementationRunner {
-    void deployToQueue();
-}
