+++ src/main/java/tdl/client/runner/QuietImplementationRunner.java
@@ -1,8 +1,8 @@
 package tdl.client.runner;
 
-public class QuietImplementationRunner implements IImplementationRunner {
+public class QuietImplementationRunner implements ImplementationRunner {
     @Override
-    public void deployToQueue() {
+    public void run() {
         // do nothing
     }
 }
