+++ src/main/java/tdl/client/runner/RoundChangesListener.java
@@ -1,0 +1,5 @@
+package tdl.client.runner;
+
+public interface RoundChangesListener {
+    void onNewRound(String roundId, String shortName);
+}
